+++ pages/debug_plot.py
@@ -1,0 +1,20 @@
+from tkinter import Frame
+
+
+class Debug_Plot(Frame):
+    def __init__(self, controller, parent):
+        super().__init__(parent)
+        self.controller = controller
+        self.name = "Debug Plot"
+
+    def create_view(self):
+        pass
+
+    def enter_button_pressed(self):
+        pass
+
+    def up_button_pressed(self):
+        pass
+
+    def down_button_pressed(self):
+        pass
